+++ src/core/pathfinding.h
@@ -15,7 +15,8 @@ namespace pathfinding {
 	std::vector<GameTile> getShortestPath(const GameBoard& board, const GameTile& source, const GameTile& destination);
 	std::vector<GameTile> _createOrderedPath(const std::unordered_map<GameTile, GameTile>&, const GameTile& source, const GameTile& destination);
 	GameTile _roundToTile(const float q, const float r, const float s);
-	inline float _lerp(const double a, const double b, const double t);
+	// linear interpolation
+	inline float _lerp(const float a, const float b, const float t);
 	std::tuple<float, float, float> _cubeLerp(const GameTile& a, const GameTile& b, const float t);
 	std::vector<GameTile> drawLine(const GameTile& source, const GameTile& destination);
 }
